+++ plugins/alignments/src/SVGErrorBox.tsx
@@ -1,14 +1,3 @@
-export function isReadyOrHasError(display: {
-  renderProps(): { notReady?: boolean }
-  error?: unknown
-}) {
-  return !display.renderProps().notReady || !!display.error
-}
-
-export function hasAnySubDisplayError(displays: Array<{ error?: unknown }>) {
-  return displays.some(d => !!d.error)
-}
-
 export function ErrorBox({
   error,
   width,
