+++ plugins/linear-genome-view/src/LinearGenomeView/svgExportUtil.ts
@@ -1,0 +1,10 @@
+export function isReadyOrHasError(display: {
+  renderProps(): { notReady?: boolean }
+  error?: unknown
+}) {
+  return !display.renderProps().notReady || !!display.error
+}
+
+export function hasAnySubDisplayError(displays: { error?: unknown }[]) {
+  return displays.some(d => !!d.error)
+}
