+++ plugins/wiggle/src/SVGErrorBox.tsx
@@ -1,0 +1,18 @@
+export function ErrorBox({
+  error,
+  width,
+  height,
+}: {
+  error: unknown
+  width: number
+  height: number
+}) {
+  return (
+    <>
+      <rect x={0} y={0} width={width} height={height} fill="#ffdddd" />
+      <text x={10} y={height / 2} fill="#cc0000" fontSize={14}>
+        {`${error}`}
+      </text>
+    </>
+  )
+}
